refactor: GsSceneLoader takes SceneLoadContext instead of AppBase&
GsSceneLoader is now fully decoupled from AppBase. It receives only the
specific subsystems it needs via SceneLoadContext, with const FeatureFlags
enforcing read-only access to feature configuration.
